Fetch TV details
