+++ pages/person/_id.vue
@@ -1,0 +1,244 @@
+<template>
+  <v-container>
+    <div v-if="hasError">
+      <h4 class="text-h4 text-center font-weight-light">{{ errorMsg }} 😖</h4>
+      <!-- <p class="subtitle-1 text-center font-weight-light">
+        Maximum requests for today 😢
+        Please come back tomorrow
+      </p> -->
+    </div>
+
+    <v-container v-else>
+      <v-row>
+        <v-col cols="12" sm="6" md="4">
+          <v-img
+            :src="imageSrc"
+            :alt="`photo of ${actor.name}`"
+            lazy-src="nopicture.jpg"
+            contain
+            width="100%"
+            aspect-ratio="2/3"
+          />
+        </v-col>
+        <v-col cols="12" sm="6" md="8" class="text-left">
+          <h2 v-if="$vuetify.breakpoint.smAndDown" class="text-h3 mb-2">
+            {{ actor.name }}
+          </h2>
+          <h2 v-else class="text-h2 mb-2">{{ actor.name }}</h2>
+          <v-divider dark class="my-2" />
+          <p>{{ actor.biography }}</p>
+          <v-divider dark class="my-2" />
+          <v-chip outlined dark color="primary" class="mr-2">
+            {{ actor.known_for_department }}
+          </v-chip>
+          <v-divider dark class="my-2" />
+          <!-- TODO: Format Date -->
+          <p>Born: {{ actor.birthday }}</p>
+          <div v-if="actor.deathday">
+            <v-divider dark class="my-2" />
+            <p>Death: {{ actor.deathday }}</p>
+          </div>
+          <v-divider dark class="my-2" />
+          <p>Born: {{ actor.place_of_birth }}</p>
+        </v-col>
+      </v-row>
+    </v-container>
+
+    <div v-if="!hasError">
+      <!-- <v-divider dark class="my-4 primary" /> -->
+
+      <!-- <h4 class="mb-2 text-h4 font-weight-light">Known For</h4> -->
+      <!-- <div class="d-flex py-4 item-container">
+        <Poster
+          v-for="kfor in actor.knownFor"
+          :key="`kfor_${kfor.id}`"
+          :imageSrc="kfor.image"
+          :name="kfor.title"
+          :role="kfor.role"
+          @on-click="onClickTitle(kfor.id)"
+        />
+      </div> -->
+
+      <v-divider dark class="my-4 primary" />
+
+      <h4 class="mb-2 text-h4 font-weight-light">Cast Credits</h4>
+
+      <v-container>
+        <v-row>
+          <v-col
+            v-for="role in actor.roles"
+            class="d-flex child-flex"
+            lg="2"
+            md="3"
+            sm="4"
+          >
+            <Poster
+              :key="role.id"
+              :imageSrc="getRoleImage(role.poster_path)"
+              :name="role.title"
+              :role="role.character"
+              @on-click="onClickTitle(role.id, role.credit_id)"
+            />
+          </v-col>
+        </v-row>
+      </v-container>
+
+      <v-divider dark class="my-4 primary" />
+
+      <h4 class="mb-2 text-h4 font-weight-light">Crew Credits</h4>
+
+      <v-container>
+        <v-row>
+          <v-col
+            v-for="credit in crew"
+            class="d-flex child-flex"
+            lg="2"
+            md="3"
+            sm="4"
+          >
+            <Poster
+              :key="credit.id"
+              :imageSrc="getRoleImage(credit.poster_path)"
+              :name="credit.title"
+              :role="credit.job"
+              @on-click="onClickTitle(role.id, role.credit_id)"
+            />
+          </v-col>
+        </v-row>
+      </v-container>
+
+      <!-- <TitleItem v-for="title in actor.roles" :key="title.id" :title="title" /> -->
+      <!-- <div
+        v-for="(title, index) in actor.roles"
+        :key="index"
+        @click="onClickTitle(title.id)"
+      >
+        <v-hover v-slot="{ hover }">
+          <div class="d-flex py-4" :class="{ 'on-hover': hover }">
+            <div>
+              <h3>{{ title.title }} {{ title.year }}</h3>
+              <p class="mb-0 subtitle-1">{{ title.role }}</p>
+              <p class="subtitle-1">{{ title.description }}</p>
+            </div>
+          </div>
+        </v-hover>
+
+        <v-divider />
+      </div> -->
+    </div>
+  </v-container>
+</template>
+
+<script>
+import mockData from "@/helpers/mockData";
+import { mapGetters, mapActions } from "vuex";
+
+import { BASE_URL, API_KEY } from "../../env";
+import Poster from "@/components/Poster.vue";
+import TitleItem from "@/components/Title-Item.vue";
+
+export default {
+  name: "ActorPage",
+  components: {
+    Poster,
+    TitleItem,
+  },
+  data() {
+    return {
+      hasMaxRequests: false,
+      movies: [],
+      hasError: false,
+      errorMsg: "",
+    };
+  },
+  computed: {
+    ...mapGetters(["actor"]),
+    imageSrc() {
+      if (this?.actor?.profile_path) {
+        return `https://image.tmdb.org/t/p/w500${this.actor.profile_path}`;
+      } else {
+        return "nopicture.jpg";
+      }
+    },
+    crew() {
+      //TODO: Fix this
+      const crew = this?.actor?.movie_credits?.crew || [];
+      const copyOfCrew = [...crew];
+      copyOfCrew.sort((a, b) => b.popularity - a.popularity);
+      return copyOfCrew;
+    },
+    // actor() {
+    //   const actorId = this.$route.params.id;
+    //   console.log("actor ~ actorId:", actorId);
+    //   const actor = this.actorById(actorId);
+    //   return actor ? actor : {};
+    // },
+  },
+  created() {
+    this.init();
+    // this.getMockData();
+  },
+  methods: {
+    ...mapActions(["fetchActor"]),
+    async init() {
+      // Check if actor is already loaded
+      const actorId = this.$route.params.id;
+      if (this.actor.id === actorId) return;
+
+      try {
+        await this.fetchActor({ actorId });
+        console.log("Fetched Actor");
+
+        console.log(this.actor);
+      } catch (err) {
+        this.hasError = true;
+        this.errorMsg = err.message;
+      }
+    },
+    getRoleImage(posterPath) {
+      if (posterPath) {
+        return `https://image.tmdb.org/t/p/w185${posterPath}`;
+      } else {
+        return "nopicture.jpg";
+      }
+    },
+    // getMockData() {
+    //   const { stars } = mockData();
+    //   this.actor = stars[1];
+    //   this.movies = this.actor.castMovies.slice(0, 10);
+    // },
+    async onClickTitle(id, credit_id) {
+      //TODO: Move to helper function
+      if (!BASE_URL || !API_KEY) throw new Error("API Key Missing");
+
+      const encodedText = encodeURIComponent(credit_id);
+      const url = `${BASE_URL}/credit/${encodedText}`;
+      const credit = await fetch(url, {
+        headers: {
+          Authorization: `Bearer ${API_KEY}`,
+        },
+      }).then(async (response) => {
+        return await response.json();
+      });
+
+      if (credit.media_type === "movie") {
+        this.$router.push(`/movie/${id}`);
+      } else if (credit.media_type === "tv") {
+        this.$router.push(`/tv/${id}`);
+      }
+      // this.$router.push(`/title/${id}`);
+    },
+  },
+};
+</script>
+
+<style scoped>
+.on-hover {
+  background-color: #272727;
+}
+
+.item-container {
+  overflow-x: auto;
+  overflow-y: hidden;
+}
+</style>
